swapper: create a new token swapping service
This change adds a simplistic token swap service which
accepts tokens to its address (signed ticket) and transfers
a new token of similar value back to the sender.

The sender is identified using the ticket they transferred
the token from.

The value handling in this service is very incomplete, as is
the token supply. This service simply provides a proof of concept.

TEST=run_servers; run_client swapper