build fixes: CMake deprecation and getentropy()
Newer versions of CMake trigger fatal errors on outdated CMake minimum required
version fields below 3.5.  While xaptum's repos should be updated (later!) this
change simply updates the minimum in both amcl and ecdaa from 3.1/3.0 to 3.5 if
CMake is new enough to have trouble.

Note, the test is not robust to a major version change in CMake.

This change also introduces an untested move from getrandom() to getentropy().
This should make the crypto portable between Linux and MacOS.  At present, this
change has only been tested on MacOS and should not be pulled until it is
tested on Linux as well.